core: Support java 16+ records with a multi-release JAR
