+++ core/src/main/java/com/electronwill/nightconfig/core/serde/ObjectSerializerBuilder.java
@@ -1,15 +1,10 @@
 package com.electronwill.nightconfig.core.serde;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
-
-import com.electronwill.nightconfig.core.Config;
 import com.electronwill.nightconfig.core.ConfigFormat;
 import com.electronwill.nightconfig.core.UnmodifiableConfig;
 
@@ -23,7 +18,7 @@ public final class ObjectSerializerBuilder {
     final List<ValueSerializerProvider<?, ?>> generalProviders = new ArrayList<>();
 
     /** the last-resort serializer provider, used when no other provider matches */
-    ValueSerializerProvider<?, ?> defaultProvider = new NoProvider();
+    ValueSerializerProvider<?, ?> defaultProvider = NoProvider.INSTANCE;
 
     /** setting: skip transient fields as requested by the modifier */
     boolean applyTransientModifier = true;
@@ -62,8 +57,8 @@ public final class ObjectSerializerBuilder {
 
     @SuppressWarnings({ "rawtypes", "unchecked" })
     public void withDefaultSerializerProvider() {
-        ValueSerializer trivialSer = new TrivialSerializer();
-        ValueSerializer fieldsSer = new FieldsSerializer();
+        ValueSerializer trivialSer = new StandardSerializers.TrivialSerializer();
+        ValueSerializer fieldsSer = new StandardSerializers.FieldsToConfigSerializer();
         ValueSerializer numberToIntSer = (value, ctx) -> ((Number) value).intValue();
         ValueSerializer charToIntSer = (value, ctx) -> (int) (Character) value;
 
@@ -71,8 +66,9 @@ public final class ObjectSerializerBuilder {
             ConfigFormat<?> format = ctx.configFormat();
             if (format == null || format.supportsType(valueClass)) {
                 return trivialSer;
-            } else if (valueClass != null && (Util.isPrimitiveOrWrapper(valueClass) || valueClass == String.class
-                    || valueClass.isArray())) {
+            } else if (valueClass != null
+                    && (Util.isPrimitiveOrWrapper(valueClass) || valueClass == String.class
+                            || valueClass.isArray())) {
                 // Cannot access the fields of the value!
                 // try to convert to int, if supported
                 if (format.supportsType(int.class)) {
@@ -102,12 +98,12 @@ public final class ObjectSerializerBuilder {
     private void registerStandardSerializers() {
         withDefaultSerializerProvider();
 
-        ValueSerializer mapSer = new MapSerializer();
-        ValueSerializer collSer = new CollectionSerializer();
-        ValueSerializer iterSer = new IterableSerializer();
-        ValueSerializer arraySer = new ArraySerializer();
-        ValueSerializer enumSer = new EnumSerializer();
-        ValueSerializer trivialSer = new TrivialSerializer();
+        ValueSerializer mapSer = new StandardSerializers.MapSerializer();
+        ValueSerializer collSer = new StandardSerializers.CollectionSerializer();
+        ValueSerializer iterSer = new StandardSerializers.IterableSerializer();
+        ValueSerializer arraySer = new StandardSerializers.ArraySerializer();
+        ValueSerializer enumSer = new StandardSerializers.EnumSerializer();
+        ValueSerializer trivialSer = new StandardSerializers.TrivialSerializer();
 
         withSerializerProvider((valueClass, ctx) -> {
             if (valueClass == null) {
@@ -140,109 +136,10 @@ public final class ObjectSerializerBuilder {
         });
     }
 
-    /**
-     * Serializes a Map<String, V> to a Config by converting each entry of the map into an entry of the config,
-     * converting the entry's value.
-     */
-    private static final class MapSerializer implements ValueSerializer<Map<?, ?>, Config> {
-
-        @Override
-        public Config serialize(Map<?, ?> value, SerializerContext ctx) {
-            Config res = ctx.createConfig();
-
-            // serialize each entry as a config entry
-            for (Entry<?, ?> entry : value.entrySet()) {
-                // get the path
-                Object key = entry.getKey();
-                if (!(key instanceof String)) {
-                    String keyTypeString = key == null ? "null" : key.getClass().toString();
-                    throw new SerializationException("Map keys must be strings, invalid key type "
-                            + keyTypeString + " in value.");
-                }
-                List<String> path = Collections.singletonList((String) key);
-
-                // convert the value
-                Object serialized = ctx.serializeValue(entry.getValue());
-
-                // add the value to the config
-                res.set(path, serialized);
-            }
-            return res;
-        }
-
-    }
-
-    private static final class CollectionSerializer implements ValueSerializer<Collection<?>, List<?>> {
-
-        @Override
-        public List<?> serialize(Collection<?> value, SerializerContext ctx) {
-            List<Object> res = new ArrayList<>(value.size());
-            for (Object v : value) {
-                Object serialized = ctx.serializeValue(v);
-                res.add(serialized);
-            }
-            return res;
-        }
-    }
-
-    private static final class IterableSerializer implements ValueSerializer<Iterable<?>, List<?>> {
-
-        @Override
-        public List<?> serialize(Iterable<?> value, SerializerContext ctx) {
-            List<Object> res = new ArrayList<>();
-            for (Object v : value) {
-                Object serialized = ctx.serializeValue(v);
-                res.add(serialized);
-            }
-            return res;
-        }
-    }
-
-    private static final class ArraySerializer implements ValueSerializer<Object, List<?>> {
-
-        @Override
-        public List<?> serialize(Object arrayValue, SerializerContext ctx) {
-            int size = Array.getLength(arrayValue);
-
-            List<Object> res = new ArrayList<>(size);
-            for (int i = 0; i < size; i++) {
-                Object element = Array.get(arrayValue, i);
-                Object serialized = ctx.serializeValue(element);
-                res.add(serialized);
-            }
-            return res;
-        }
-    }
-
-    private static final class EnumSerializer implements ValueSerializer<Enum<?>, String> {
-
-        @Override
-        public String serialize(Enum<?> value, SerializerContext ctx) {
-            return value.name();
-        }
-    }
-
-    private static final class TrivialSerializer implements ValueSerializer<Object, Object> {
-
-        @Override
-        public Object serialize(Object value, SerializerContext ctx) {
-            return value;
-        }
-    }
-
-    /** Converts an object to a config by turning each field into an entry. */
-    private static final class FieldsSerializer implements ValueSerializer<Object, Config> {
-
-        @Override
-        public Config serialize(Object value, SerializerContext ctx) {
-            Config sub = ctx.createConfig();
-            ctx.serializeFields(value, sub);
-            return sub;
-        }
-    }
-
     /** A provider that provides nothing, {@code provide} always returns null. */
-    private static final class NoProvider implements ValueSerializerProvider<Object, Object> {
+    static final class NoProvider implements ValueSerializerProvider<Object, Object> {
+        static final NoProvider INSTANCE = new NoProvider();
+
         @Override
         public ValueSerializer<Object, Object> provide(Class<?> valueClass, SerializerContext ctx) {
             return null;
