+++ examples/build.gradle
@@ -5,10 +5,6 @@ description "$projectDescription - $name module"
 apply plugin: 'java'
 apply plugin: 'application'
 
-java {
-	sourceCompatibility = JavaVersion.VERSION_1_8
-}
-
 repositories {
     mavenCentral()
 }
